% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_2st.R
\name{al_2st}
\alias{al_2st}
\title{Assurance Level for Independent Sample t-tests.}
\usage{
al_2st(
  n1,
  n2,
  d_est,
  d_sd,
  alpha = 0.05,
  power = 0.8,
  prior_d = c("norm", "trunc_norm"),
  trunc_d = c(-Inf, Inf),
  ndraws = 1e+06,
  test = "two.sided"
)
}
\arguments{
\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{d_est}{Effect size estimate, defined as
\eqn{d = \frac{\bar x_2 - \bar x_1}
  {\sqrt{\frac{(n_1 - 1)s_1^2 + (n_2 - 1)s_2^2}{n_1 + n_2 - 2}}}}}

\item{d_sd}{Uncertainty level of the effect size estimate.}

\item{alpha}{Type I error rate. Default to be \code{.05}.}

\item{power}{Desired statistical power to achieve. Default to be \code{.8}.}

\item{test}{One-sided or two-sided test. Options are either "one.sided" or "two.sided".}
}
\value{
The assurance level given the sample size and the prior for the effect size.
}
\description{
\code{al_crt()} computes the assurance level of power given the estimates and
the uncertainty level of the parameter estimates for an independent sample t-test.
}
\examples{
al_2st(n1 = 100, n2 = 100, d_est = .4, d_sd = .2)
}
