% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_indp_t.R
\name{al_indp_t}
\alias{al_indp_t}
\title{Determine Assurance Level (AL) for Independent Sample T-Tests}
\usage{
al_indp_t(
  n1,
  n2,
  delta,
  delta_sd,
  alpha = 0.05,
  power = 0.8,
  test = "two.sided"
)
}
\arguments{
\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{delta}{Effect size estimate, defined as
\eqn{d = \frac{\bar x_2 - \bar x_1}
  {\sqrt{\frac{(n_1 - 1)s_1^2 + (n_2 - 1)s_2^2}{n_1 + n_2 - 2}}}}}

\item{delta_sd}{Uncertainty level of the effect size estimate.}

\item{alpha}{Type I error rate. Defaults to \code{.05}.}

\item{power}{Desired level of statistical power. Defaults to \code{.8}.}

\item{test}{One-sided or two-sided test. Defaults to "two.sided".}
}
\value{
Assurance level given the sample size and the prior for the effect size.
}
\description{
\code{al_indp_t()} computes the assurance level of power given the effect size
estimates and the uncertainty level of the parameter estimates for an
independent sample t-test.
}
\examples{
al_indp_t(n1 = 100, n2 = 100, delta = .4, delta_sd = .2)
}
