% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_ab.R
\name{beta_ab}
\alias{beta_ab}
\title{Shape Parameters of a Beta Distribution}
\usage{
beta_ab(mode, sd)
}
\arguments{
\item{mode}{Mode of the parameter.}

\item{sd}{Standard deviation of the parameter.}
}
\value{
Shape hyperparameters of a beta distribution.
}
\description{
\code{beta_ab()} computes the shape hyperparameters of a beta distribution
from the mode and standard deviation of a parameter (e.g., intraclass correlation).
}
\examples{
beta_ab(.1, .05)
beta_ab(.5, .2)
}
