% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_hcbr.R
\name{compute_icc}
\alias{compute_icc}
\title{Compute intraclass correlation (ICC)}
\usage{
compute_icc(r_sq, sigma_sq)
}
\arguments{
\item{r_sq}{The population between-group variance.}

\item{sigma_sq}{The population within-group variance.}
}
\value{
The ICC value, a number between 0 and 1.
}
\description{
\code{compute_icc()} computes ICC as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}.
}
