% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_hcbr.R
\name{define_loss}
\alias{define_loss}
\title{Define loss function for optimization}
\usage{
define_loss(
  solve_for_J = TRUE,
  squared = FALSE,
  ep,
  al,
  J,
  n,
  test,
  reparameterize,
  list_params
)
}
\arguments{
\item{solve_for_J}{Indicate whether to define for the loss function for J.
If FALSE, define it for n.}

\item{squared}{Indicate whether the loss function should be squared. If
FALSE, set power_of to 1.}
}
\description{
Define loss function for optimization
}
