% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_crt2.R
\name{ep_crt2}
\alias{ep_crt2}
\title{Expected Power for Two-Level CRTs}
\usage{
ep_crt2(
  J,
  n,
  delta,
  delta_sd,
  rho,
  rho_sd,
  rsq2 = 0,
  K = 0,
  P = 0.5,
  power = 0.8,
  alpha = 0.05,
  test = "two.sided",
  minEval = 50,
  reparameterize = FALSE,
  ...
)
}
\arguments{
\item{J}{Number of clusters. Determine \code{n} if \code{J} is specified.}

\item{n}{Cluster size. Determine \code{J} if \code{n} is specified.}

\item{delta}{Effect size estimate, defined as
\eqn{\delta = \frac{\gamma_{01}}{\tau^2 + \sigma^2}},
where \eqn{\gamma_{01}} is the main effect of the treatment on the outcome,
\eqn{\tau^2} is the variance of the cluster-specific random effect
in the unconditional model (without covariates), and
\eqn{\sigma^2} is the variance of the random error in the unconditional model.}

\item{delta_sd}{Uncertainty level of the effect size estimate.}

\item{rho}{Intraclass correlation estimate, defined as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}, where \eqn{\tau^2} and \eqn{\sigma^2}
are the variance components in the unconditional model.}

\item{rho_sd}{Uncertainty level of the intraclass correlation estimate.}

\item{rsq2}{Estimate of variance explained by the cluster-level covariates.}

\item{K}{Number of cluster-level covariates.}

\item{P}{Proportion of the clusters that is treatment group.}

\item{power}{Desired statistical power to achieve. Default to be \code{.8}.}

\item{alpha}{Type I error rate. Default to be \code{.05}.}

\item{test}{One-sided or two-sided test. Options are either "one.sided" or "two.sided".}

\item{minEval}{Minimum number of evaluations of the integral.
Defaults to \code{50} to stabilize results.}

\item{...}{Additional arguments to be passed to \code{cubuture::cuhre()} to evaluate
the integral.}
}
\value{
The expected power given a two-level CRT design with J clusters each
with n observations.
}
\description{
\code{ep_crt2()} computes the expected power over the specified uncertainty
about the parameters for a two-level CRT design.
}
\examples{
ep_crt2(J = 30, n = 100, delta = .5, delta_sd = .2, rho = .1, rho_sd = .05)
ep_crt2(J = 30, n = 100, delta = .5, delta_sd = .2, rho = .1, rho_sd = .05,
        rsq2 = .3)
}
