% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_indp_t.R
\name{ep_indp_t}
\alias{ep_indp_t}
\title{Determine Expected Power (EP) for Independent Sample T-Tests}
\usage{
ep_indp_t(
  delta,
  delta_sd,
  n1,
  n2,
  alpha = 0.05,
  power = 0.8,
  test = "two.sided"
)
}
\arguments{
\item{delta}{Effect size estimate, defined as
\eqn{d = \frac{\bar x_2 - \bar x_1}
  {\sqrt{\frac{(n_1 - 1)s_1^2 + (n_2 - 1)s_2^2}{n_1 + n_2 - 2}}}}}

\item{delta_sd}{Uncertainty level of the effect size estimate.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{alpha}{Type I error rate. Defaults to \code{.05}.}

\item{power}{Desired statistical power to achieve. Defaults to \code{.8}.}

\item{test}{One-sided or two-sided test. Defaults to "two.sided".}
}
\value{
Expected power given the sample size and the prior for the effect size.
}
\description{
\code{ep_indp_t()} computes the expected power given the effect size estimate
and uncertainty in the estimate for an independent sample t-test.
}
\examples{
ep_indp_t(n1 = 100, n2 = 100, delta = .4, delta_sd = .2)
}
