% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ese_msrt2.R
\name{ese_msrt2}
\alias{ese_msrt2}
\title{Expected Confidence Interval Half Width for Two-Level Multisite Randomized Trials}
\usage{
ese_msrt2(
  rho,
  rho_sd,
  omega,
  omega_sd,
  J,
  n,
  se = 0.05,
  rsq1 = 0,
  rsq2 = 0,
  K = 0,
  P = 0.5,
  ...
)
}
\arguments{
\item{rho}{Intraclass correlation value, defined as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}, where \eqn{\tau^2} and \eqn{\sigma^2}
are the variance components in the unconditional model.}

\item{rho_sd}{Uncertainty level of the intraclass correlation value.}

\item{omega}{Treatment effect hetereogeneity value, defined as
\eqn{\omega = \frac{\tau_1^2}{\tau_0^2}} where \eqn{\tau_0^2} is the variance of the
intercept random component and \eqn{\tau_1^2} is the variance of the treatment
random effect.}

\item{omega_sd}{Uncertainty level of the treatment effect hetereogeneity value.}

\item{J}{Number of clusters. Determine \code{n} if \code{J} is specified.}

\item{n}{Cluster size. Determine \code{J} if \code{n} is specified.}

\item{se}{Desired confidence interval half width (standard error) to achieve.}

\item{rsq1}{Estimate of variance explained by the level-1 (e.g., individual-level) covariates.}

\item{rsq2}{Estimate of variance explained by the cluster-level covariates.}

\item{K}{Number of cluster-level covariates.}

\item{P}{Proportion of the clusters that is treatment group.}
}
\value{
The expected confidence interval half width for two-level mulsite randomized trials.
}
\description{
Expected Confidence Interval Half Width for Two-Level Multisite Randomized Trials
}
