% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_ab.R
\name{gamma_ab}
\alias{gamma_ab}
\title{Shape and scale hyperparameters of a Gamma Distribution}
\usage{
gamma_ab(mode, sd)
}
\arguments{
\item{mode}{Mode of the parameter.}

\item{sd}{Standard deviation of the parameter.}
}
\value{
Shape and scale hyperparameters of a gamma distribution.
}
\description{
\code{gamma_ab()} computes the shape and scale hyperparameters of a gamma
distribution from the mode and standard deviation of a parameter (e.g.,
effect size heterogeneity).
}
\examples{
gamma_ab(.1, .05)
gamma_ab(.5, .2)
}
