% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ab.R
\name{get_ab}
\alias{get_ab}
\title{Shape Parameters of a Beta Distribution}
\usage{
get_ab(est, sd)
}
\arguments{
\item{est}{Intraclass correlation estimate.}

\item{sd}{Uncertainty in the intraclass correlation estimate.}
}
\value{
Shape paraemters for a beta distribution given the mean and standard deviation.
}
\description{
\code{get_ab()} computes the shape parameters of a beta distribution from the mean
and standard deviation of a parameter.
}
\examples{
get_ab(.1, .05)
get_ab(.5, .2)
}
