% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcbApp.R
\name{hcbApp}
\alias{hcbApp}
\title{An Interactive Shiny App for HCB Power Analysis}
\usage{
hcbApp()
}
\value{
An interactive Shiny App to perform HCB power analysis.
}
\description{
\code{hcbApp()} executes the interactive Shiny App for HCB power analysis.
}
