% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_hcbr.R
\name{Jn_msrt2_c}
\alias{Jn_msrt2_c}
\title{\code{Jn_msrt2_c()} solves J, n using the conventional approach for a
level-two MSRT.}
\usage{
Jn_msrt2_c(
  delta,
  rho,
  omega,
  rsq1 = 0,
  rsq2 = 0,
  J = NULL,
  n = NULL,
  K = 0,
  P = 0.5,
  alpha = 0.05,
  power = 0.8,
  test = "two.sided"
)
}
\arguments{
\item{delta}{Effect size estimate.}

\item{rho}{Intraclass correlation (ICC) estimate.}

\item{omega}{Estimate of the treatment effect heterogeneity.}

\item{rsq1}{Estimate of variance explained by the level-1 (individual-level)
covariates.}

\item{rsq2}{Estimate of variance explained by the level-2 (cluster-level)
covariates.}

\item{J}{Number of clusters. If \code{J} is specified, \code{Jn_msrt2()}
determines \code{n}.}

\item{n}{Cluster size. If \code{n} is specified, \code{Jn_msrt2()}
determines \code{J}.}

\item{K}{Number of cluster-level covariates. \code{0} by default.}

\item{P}{Proportion of the clusters that is treatment group. \code{.5} by
default.}

\item{alpha}{Type I error rate. \code{.05} by default.}

\item{power}{Desired level of statistical power. \code{.8} by default.}

\item{test}{Whether a one-sided or two-sided test should be performed.
Defaults to "two-sided".}
}
\value{
A 1 x 2 array containing the J and n values.
}
\description{
\code{Jn_msrt2_c()} solves J, n using the conventional approach for a
level-two MSRT.
}
