% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jn_msrt2_draws.R
\name{Jn_msrt2_draws}
\alias{Jn_msrt2_draws}
\title{Determine Number of Clusters or Cluster Size for Two-Level Multisite Randomzied Trials}
\usage{
Jn_msrt2_draws(
  draws_d,
  draws_rho,
  draws_omega,
  rsq1 = 0,
  rsq2 = 0,
  J = NULL,
  n = NULL,
  K = 0,
  P = 0.5,
  alpha = 0.05,
  power = 0.8,
  precision = 0.15,
  ep = NULL,
  al = NULL,
  apr = NULL,
  test = "two.sided",
  plot = FALSE
)
}
\description{
Determine Number of Clusters or Cluster Size for Two-Level Multisite Randomzied Trials
}
