% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jn_msrt2_prec.R
\name{Jn_msrt2_prec}
\alias{Jn_msrt2_prec}
\title{Determine Number of Clusters or Cluster Size for Two-Level Multisite Randomized Trials}
\usage{
Jn_msrt2_prec(
  rho,
  rho_sd,
  omega,
  omega_sd,
  rsq1 = 0,
  rsq2 = 0,
  J = NULL,
  n = NULL,
  K = 0,
  P = 0.5,
  precision = 0.1,
  apr = 0.6,
  plot = FALSE
)
}
\arguments{
\item{rho}{Intraclass correlation value, defined as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}, where \eqn{\tau^2} and \eqn{\sigma^2}
are the variance components in the unconditional model.}

\item{rho_sd}{Uncertainty level of the intraclass correlation value.}

\item{omega}{Treatment effect hetereogeneity value, defined as
\eqn{\omega = \frac{\tau_1^2}{\tau_0^2}} where \eqn{\tau_0^2} is the variance of the
intercept random component and \eqn{\tau_1^2} is the variance of the treatment
random effect.}

\item{omega_sd}{Uncertainty level of the treatment effect hetereogeneity value.}

\item{rsq1}{Estimate of variance explained by the level-1 (e.g., individual-level) covariates.}

\item{rsq2}{Estimate of variance explained by the cluster-level covariates.}

\item{J}{Number of clusters. Determine \code{n} if \code{J} is specified.}

\item{n}{Cluster size. Determine \code{J} if \code{n} is specified.}

\item{K}{Number of cluster-level covariates.}

\item{P}{Proportion of the clusters that is treatment group.}

\item{plot}{If TRUE, plots of J and n against the expected power or assurance level
will be printed.}

\item{ese}{Desired expected confidence interval half width to achieve.
If neither \code{ese} nor \code{ase} is specified, \code{ese} = \code{se}.
An 0.05 expected half width indicates that the average confidence interval
half width is 0.05 over the specified uncertainty.}

\item{ase}{Assurance level of confidence interval half width to achieve.
An 80\% assurance level indicates there is an 80\% chance that the confidence interval
half width is below or at the desired half width over the specified uncertainty.}
}
\value{
The required J or n and optional plots that show the curves of
expected power/assurance level.
}
\description{
Determine Number of Clusters or Cluster Size for Two-Level Multisite Randomized Trials
}
\seealso{
\url{https://winnie-wy-tse.shinyapps.io/hcb_shiny/}
}
