% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_indp_t.R
\name{n_indp_t}
\alias{n_indp_t}
\title{Determine Cluster Size (n) for Independent Sample T-Tests}
\usage{
n_indp_t(
  delta,
  delta_sd,
  alpha = 0.05,
  power = 0.8,
  ep = NULL,
  al = NULL,
  test = "two.sided",
  plot = FALSE
)
}
\arguments{
\item{delta}{Effect size estimate, defined as
\eqn{d = \frac{\bar x_2 - \bar x_1}
  {\sqrt{\frac{(n_1 - 1)s_1^2 + (n_2 - 1)s_2^2}{n_1 + n_2 - 2}}}}}

\item{delta_sd}{Uncertainty level of the effect size estimate.}

\item{alpha}{Type I error rate. Defaults to \code{.05}.}

\item{power}{Desired statistical power level.}

\item{ep}{Desired expected power (EP) level. For example, an 80\% EP
indicates that the mean or average power value is 80\% over the specified
uncertainty. If neither \code{ep} nor \code{al} is specified,
\code{ep} = \code{power}.}

\item{al}{Assurance level (AL) to achieve. For example, an 80\% AL indicates
that 80\% of the power values are above the desired statistical power over
the specified uncertainty.}

\item{test}{One-sided or two-sided test. Defaults to "two-sided".}

\item{plot}{If TRUE, plots of J and n against EP or AL will be printed.}
}
\value{
Required group size for a two-sample t-test.
}
\description{
\code{n_indp_t()} solves for the required cluster size for a two-sample t-test.
When the uncertainty level of the effect size is specified, determines the
sample size requisite that achieves the desired expected power or assurance
level. Otherwise, determines the sample size requisite that achieves the
desired classical power.
}
\examples{
n_indp_t(delta = .5, delta_sd = .1)
}
\seealso{
\url{https://winnie-wy-tse.shinyapps.io/hcb_shiny/}
}
