% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Prior Distributions of Parameters}
\usage{
plot_prior(d_est, d_sd, rho_est, rho_sd)
}
\arguments{
\item{d_est}{Effect size estimate, defined as
\eqn{\delta = \frac{\gamma_{01}}{\tau^2 + \sigma^2}},
where \eqn{\gamma_{01}} is the main effect of the treatment on the outcome,
\eqn{\tau^2} is the variance of the cluster-specific random effect
in the unconditional model (without covariates), and
\eqn{\sigma^2} is the variance of the random error in the unconditional model.}

\item{d_sd}{Uncertainty level of the effect size estimate.}

\item{rho_est}{Intraclass correlation estimate, defined as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}, where \eqn{\tau^2} and \eqn{\sigma^2}
are the variance components in the unconditional model.}

\item{rho_sd}{Uncertainty level of the intraclass correlation estimate.}
}
\description{
\code{plot_prior()} generate plots of the prior distributions of effect size
and/or intraclass correlation.
}
