% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_2st.R
\name{pow_2st}
\alias{pow_2st}
\title{Statistical Power for Independent Sample t-tests}
\usage{
pow_2st(d_est, n1, n2, alpha = 0.05, test = "two.sided")
}
\arguments{
\item{d_est}{Effect size estimate.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{alpha}{Type I error rate. Default to be \code{.05}.}

\item{test}{One-tailed or two-tailed test.}
}
\value{
Statistical power given the sample size for an independent sample t-test.
}
\description{
\code{pow_2st()} computes the statistical power for an independent sample t-test.
}
\examples{
pow_2st(n1 = 100, n2 = 100, d_est = .4)
}
