% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_crt2.R
\name{pow_crt2}
\alias{pow_crt2}
\title{Statistical Power for Two-Level CRTs}
\usage{
pow_crt2(J, n, d_est, rho_est, r2_est = 0, K = 0, P = 0.5, test = "two-tailed")
}
\arguments{
\item{J}{Specified number of clusters.}

\item{n}{Specified cluster size.}

\item{d_est}{Effect size estimate.}

\item{rho_est}{Intraclass correlation estimate.}

\item{r2_est}{Estimate of variance explained by the cluster-level covariates.}

\item{K}{Number of cluster-level covariates.}

\item{test}{One-tailed or two-tailed test.}
}
\value{
Statistical power given certain J and n for cluster-randomized trials.
}
\description{
\code{pow_crt2()} computes the statistical power for a two-level CRT.
}
\examples{
pow_crt2(J = 30, n = 100, d_est = .5, rho_est = .1)
pow_crt2(J = 30, n = 100, d_est = .5, rho_est = .1, r2_est = .3)
}
