% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_indp_t.R
\name{pow_indp_t}
\alias{pow_indp_t}
\title{Determine Statistical Power for Independent Sample T-Tests}
\usage{
pow_indp_t(delta, n1, n2, alpha = 0.05, test = "two.sided")
}
\arguments{
\item{delta}{Effect size estimate.}

\item{n1}{Sample size of group 1.}

\item{n2}{Sample size of group 2.}

\item{alpha}{Type I error rate. Defaults to \code{.05}.}

\item{test}{One-tailed or two-tailed test. Defaults to "two.sided".}
}
\value{
Statistical power given the sample size for an independent sample t-test.
}
\description{
\code{pow_indp_t()} computes the statistical power for an independent sample
t-test.
}
\examples{
pow_indp_t(n1 = 100, n2 = 100, delta = .4)
}
