% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_crt.R
\name{power_crt}
\alias{power_crt}
\title{Statistical Power for Two-Level CRTs}
\usage{
power_crt(J, n, d_est, rho_est, r2_est = 0, K = 0, test = "two-tailed")
}
\arguments{
\item{J}{Specified number of clusters.}

\item{n}{Specified cluster size.}

\item{d_est}{Effect size estimate.}

\item{rho_est}{Intraclass correlation estimate.}

\item{r2_est}{Estimate of variance explained by the cluster-level covariates.}

\item{K}{Number of cluster-level covariates.}

\item{test}{One-tailed or two-tailed test.}
}
\value{
Statistical power given certain J and n for cluster-randomized trials.
}
\description{
\code{power_crt()} computes the statistical power for a two-level CRT.
}
\examples{
power_crt(J = 30, n = 100, d_est = .5, rho_est = .1)
power_crt(J = 30, n = 100, d_est = .5, rho_est = .1, r2_est = .3)
}
