% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_msrt2.R
\name{prec_msrt2}
\alias{prec_msrt2}
\title{Precision for Two-Level Multisite Randomized Trials}
\usage{
prec_msrt2(rho, omega, J, n, rsq1 = 0, rsq2 = 0, K = 0, P = 0.5, alpha = 0.05)
}
\arguments{
\item{rho}{Intraclass correlation value, defined as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}, where \eqn{\tau^2} and \eqn{\sigma^2}
are the variance components in the unconditional model.}

\item{omega}{Treatment effect hetereogeneity value, defined as
\eqn{\omega = \frac{\tau_1^2}{\tau_0^2}} where \eqn{\tau_0^2} is the variance of the
intercept random component and \eqn{\tau_1^2} is the variance of the treatment
random effect.}

\item{J}{Number of clusters. Determine \code{n} if \code{J} is specified.}

\item{n}{Cluster size. Determine \code{J} if \code{n} is specified.}

\item{rsq1}{Estimate of variance explained by the level-1 (e.g., individual-level) covariates.}

\item{rsq2}{Estimate of variance explained by the cluster-level covariates.}

\item{K}{Number of cluster-level covariates.}

\item{P}{Proportion of the clusters that is treatment group.}
}
\value{
The expected confidence interval half width for two-level multisite randomized trials.
}
\description{
Precision is defined as the confidence interval half width.
}
