% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_crt2.R
\name{ep_crt2}
\alias{ep_crt2}
\title{Expected Power for Two-Level CRTs}
\usage{
ep_crt2(
  J,
  n,
  d_est,
  d_sd,
  rho_est,
  rho_sd,
  r2_est = 0,
  r2_sd = 0,
  K = 0,
  P = 0.5,
  test = "two-tailed",
  abs.tol = 1e-50,
  rel.tol = 0.001,
  minEval = 50
)
}
\arguments{
\item{J}{Specified number of clusters.}

\item{n}{Specified cluster size.}

\item{d_est}{Effect size estimate.}

\item{d_sd}{Uncertainty in the effect size estimate.}

\item{rho_est}{Intraclass correlation estimate.}

\item{rho_sd}{Uncertainty in the intraclass correlation estimate.}

\item{r2_est}{Estimate of variance explained by the cluster-level covariates.}

\item{r2_sd}{Uncertainty in the variance explained by the cluster-level covariates.}

\item{K}{Number of cluster-level covariates.}

\item{test}{One-tailed or two-tailed test.}

\item{abs.tol}{Absolute tolerance. Defaults to \code{1e-10}.}

\item{rel.tol}{Relative tolerance. Defaults to \code{1e-15}.}

\item{minEval}{Minimum number of evaluations of the integral. Defaults to \code{50}.}
}
\value{
The expected power given certain J and n.
}
\description{
\code{ep_crt2()} computes the expected power given the estimates of and the uncertainty
in the parameter estimates for a two-level CRT.
}
\examples{
ep_crt2(J = 30, n = 100, d_est = .5, d_sd = .2, rho_est = .1, rho_sd = .05)
ep_crt2(30, 100, .5, .2, .1, .05, .3, .1)
}
