% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ab.R
\name{get_ab}
\alias{get_ab}
\title{Shape Parameters of a Beta Distribution}
\usage{
get_ab(mode, sd)
}
\arguments{
\item{mode}{Mode of the parameter.}

\item{sd}{Standard deviation of the parameter.}
}
\value{
Shape hyperparaemters of a beta distribution.
}
\description{
\code{get_ab()} computes the shape hyperparameters of a beta distribution
from the mode and standard deviation of a parameter (e.g., intraclass correlation).
}
\examples{
get_ab(.1, .05)
get_ab(.5, .2)
}
