% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jn_crt2.R
\name{Jn_crt2}
\alias{Jn_crt2}
\title{Determine Required Number of Clusters or Cluster Size for Two-Level CRTs}
\usage{
Jn_crt2(
  d_est,
  d_sd,
  rho_est,
  rho_sd,
  r2_est = 0,
  r2_sd = 0,
  J = NULL,
  n = NULL,
  K = 0,
  power = 0.8,
  al = NULL,
  test = "two-tailed",
  plot = FALSE,
  abs.tol = 1e-10,
  x.tol = 1.5e-15,
  rel.tol = 1e-15,
  sing.tol = 1e-20
)
}
\arguments{
\item{d_est}{Effect size estimate.}

\item{d_sd}{Uncertainty in the effect size estimate.}

\item{rho_est}{Intraclass correlation estimate.}

\item{rho_sd}{Uncertainty in the intraclass correlation estimate.}

\item{r2_est}{Estimate of variance explained by the cluster-level covariates.}

\item{r2_sd}{Uncertainty in the variance explained by the cluster-level covariates.}

\item{J}{Specified number of clusters.}

\item{n}{Specified cluster size.}

\item{K}{Number of cluster-level covariates.}

\item{power}{Desired power level to achieve.}

\item{test}{One-tailed or two-tailed test.}

\item{plot}{Printing out a plot if it is TURE.}

\item{abs.tol}{Absolute tolerance. Defaults to \code{1e-10}.}

\item{x.tol}{X tolerance. Defaults to \code{1.5e-15}.}

\item{rel.tol}{Relative tolerance. Defaults to \code{1e-15}.}

\item{sing.tol}{Singular convergence tolerance. Defaults to \code{1e-20}.}

\item{criteria}{Solve the required J or n based on the desired assurance level
or expected power}
}
\value{
The required J or n and a optionally plot that shows the power curve.
}
\description{
\code{Jn_crt2()} solves for the required number of clusters (J) or cluster size (n)
for a two-level CRT using the HCB approach. When certain uncertainty is specified,
this function determines the minimum required sample size that achieves the
desired level of expected power.
}
\examples{
Jn_crt2(d_est = .5, d_sd = .2, rho_est = .1, rho_sd = .05, J = 30)
}
\seealso{
\url{https://winnie-wy-tse.shinyapps.io/hcb_shiny/}
}
