% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jn_crt2.R
\name{Jn_crt2}
\alias{Jn_crt2}
\title{Determine Number of Clusters or Cluster Size for Two-Level CRTs}
\usage{
Jn_crt2(
  delta,
  delta_sd,
  rho,
  rho_sd,
  rsq2 = 0,
  J = NULL,
  n = NULL,
  K = 0,
  P = 0.5,
  alpha = 0.05,
  power = 0.8,
  ep = NULL,
  al = NULL,
  test = "two.sided",
  plot = FALSE,
  max_try = 1e+06
)
}
\arguments{
\item{delta}{Effect size estimate, defined as
\eqn{\delta = \frac{\gamma_{01}}{\tau^2 + \sigma^2}}, where
\eqn{\gamma_{01}} is the main effect of the treatment on the outcome,
\eqn{\tau^2} is the variance of the cluster-specific random effect in the
unconditional model (without covariates), and \eqn{\sigma^2} is the
variance of the random error in the unconditional model.}

\item{delta_sd}{Uncertainty level of the effect size estimate.}

\item{rho}{Intraclass correlation (ICC) estimate, defined as
\eqn{\rho = \frac{\tau^2}{\tau^2 + \sigma^2}}, where \eqn{\tau^2} and
\eqn{\sigma^2} are the variance components in the unconditional model.}

\item{rho_sd}{Uncertainty level of the ICC estimate.}

\item{rsq2}{Estimate of variance explained by the cluster-level covariates.
\code{0} by default.}

\item{J}{Number of clusters. If \code{J} is specified, \code{Jn_crt2()}
determines \code{n} for the specified expected power/assurance level.}

\item{n}{Cluster size. If \code{n} is specified, \code{Jn_crt2()}
determines \code{J} for the specified expected power/assurance level.}

\item{K}{Number of cluster-level covariates. \code{0} by default.}

\item{P}{Proportion of the clusters that are treatment groups. \code{.5} by
default.}

\item{alpha}{Type I error rate. \code{.05} by default.}

\item{power}{Desired level of statistical power. \code{.8} by default.}

\item{ep}{Desired expected power (EP). If neither \code{ep} nor \code{al} is
specified, \code{ep} = \code{power}. For example, an 80\% EP indicates that
the mean or average power value is 80\% over the specified uncertainty.}

\item{al}{Desired assurance level (AL). For example, an 80\% AL indicates 80\%
of the power values are above the desired statistical power over the
specified uncertainty.}

\item{test}{Whether a one-sided or two-sided test should be performed.}

\item{plot}{(optional) whether plots of J and n against expected power or
assurance level should be returned.}
}
\value{
A 1 x 2 array containing the J and n values that together achieve the
desired expected power or assurance level. If \code{plot = TRUE}, also
returns plots.
}
\description{
\code{Jn_crt2()} uses the HCB approach to solve for the minimum required
number of clusters (J) or cluster size (n) that would achieve a desired
expected power or assurance level for a two-level CRT.
}
\examples{
Jn_crt2(delta = .5, delta_sd = .2, rho = .1, rho_sd = .05, J = 30)
}
\seealso{
\url{https://winnie-wy-tse.shinyapps.io/hcb_shiny/}
}
